[update] Add coinone V2 api support
Some of ccxt apis, for example create_order(), for coinone didn't support V2. When an user tries to use it, it returns an error since V1 apis are NOT supported anymore by coinone. This modification corrected it.

Please be aware that we should move to V2.1 someday, since V2 is deprecated already.